Add useNavigation hook
